+++ src/Lykke.Service.HFT.Services/Consumers/Messages/LimitOrderMessage.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Lykke.Service.HFT.Core.Domain;
 
-namespace Lykke.Service.HFT.Services.Messages
+namespace Lykke.Service.HFT.Services.Consumers.Messages
 {
     public class LimitOrderMessage
     {
